requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
